requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
flask-cors==5.0.1
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
pytest==8.3.5
python-dotenv==1.0.1
requests==2.32.3
requests-mock==1.12.1
SQLAlchemy==2.0.39
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
